docs(faz8b): PR #50 review fixes — claim doğruluğu + counter u64 + INV-C11 envanteri
İki review turundan gelen düzeltmeler (request changes → resolve).

R1 P1 BLOKLAYICI (verification claim): mod.rs "no external caller can drive the
transition" yanlıştı — SupersedeSession public olduğu için external caller transition'ı
isteyebilir (INV-C11 deployment boundary). Claim structural confinement'a indirgendi:
external caller authority mint edemez/application construct edemez, ama transition'ı
SupersedeSession üzerinden isteyebilir; operator authorization INV-C11 deployment/runtime.

R1/R2 counter u32→u64 (plan kararı + audit_seq simetrisi): kod u32, doc'lar u64 diyordu —
üç yüzey üç şey söylüyordu. SupersedeSession.supersedes + SupersedeSessionSummary.supersedes
u64'e; exhaustion test sabiti u64::MAX; checked_add zaten runtime güvenliği sağlıyordu.

R2 exhaustion test iç monolog: üç paragraf çıkmaz-sokak anlatısı → tek cümle ("Test modülü
aynı dosyada olduğundan private field'a doğrudan erişilir"). PR #49 cycle-yorum dersinin
tekrarı; HANDOFF checklist'ine "commit öncesi test yorumlarında karar-öncesi anlatı taraması"
maddesi eklenebilir.

R1 scorer yorumu scope: "successor relation established atomically" fazla genişti — yalnız
apply_supersede ile transition edilen node'lar için doğru; seeded/deserialized node'lar
future persisted-graph validator gerektirir. Makale INV-C15 paragrafının "transition-generated"
scope'ına indirgendi.

R2 INV-C11 envanteri (paper 279, 296): PR protokole ikinci kapı ekledi ama makale tehdit
bölümü onu saymıyordu. "invoking OperatorReviewSession" → "OperatorReviewSession/SupersedeSession";
"every promotion carries a DecisionRecord" → "+ every supersession carries a SupersedeRecord";
constructor exposure → constructors.

R1 doğrulama dürüstlüğü: PR gövdesi "tam workspace clippy + hard format" iddia etmişti ama
gerçekte --lib + rustfmt --check (5 dosya) çalıştırılmıştı. Çalıştırıldı: workspace clippy
(44 uyarı, 43 baseline + 1 review.rs:663 pre-existing too_many_arguments PR #49; benim
eklediğim SupersedeSession/issuer/test kodunda 0 yeni uyarı); cargo fmt --all --check
(benim 5 anchoring src dosyam temiz; pre-existing diffler paper3_evidence/heldout/g2c/desktop'ta,
CI || true warning-only).

R1 test count metodolojisi: tek "762 total" yerine kapsamlı döküm (PR #49 754 vs PR #50 762
tutarsızlığı ders). HANDOFF/STATUS: osp-core lib 502, compile-fail 24, workspace 764 (osp-desktop
hariç), yeni SupersedeSession 10.

Doğrulama: 502 lib + 24 compile-fail yeşil; workspace 764 passed (osp-desktop hariç); 0 yeni
clippy; benim 5 dosyam fmt temiz.

+++ crates/osp-core/src/anchoring/review.rs
@@ -799,7 +799,7 @@ pub(crate) fn supersede_basis_fingerprint(basis: &PresentedSupersedeBasis) -> [u
 pub struct SupersedeSessionSummary {
     pub session_id: SessionId,
     pub operator: OperatorId,
-    pub supersedes: u32,
+    pub supersedes: u64,
 }
 
 /// Operator supersede oturumu — INV-C15 production invocation boundary'si.
@@ -826,7 +826,7 @@ pub struct SupersedeSession {
     session_id: SessionId,
     operator: OperatorId,
     opened_at: SystemTime,
-    supersedes: u32,
+    supersedes: u64,
 }
 
 impl SupersedeSession {
@@ -2476,7 +2476,7 @@ mod tests {
     }
 
     /// Counter exhaustion: supersedes=u64::MAX set → SessionCounterExhausted + store unchanged.
-    /// (Counter u32 → MAX set; checked_add overflow mutation öncesi yakalanır.)
+    /// checked_add overflow mutation öncesi yakalanır; counter yalnız başarılı store op sonrası assign edilir.
     #[test]
     fn supersede_session_counter_exhaustion_leaves_store_unchanged() {
         let mut store = store_with_two_accepted("RuleCandidate:Old", "RuleCandidate:New");
@@ -2485,19 +2485,8 @@ mod tests {
         let (mut session, basis) = session_with_basis(&store, &old, &new);
         let reason = NonEmptyExplanation::new("exhausted").unwrap();
 
-        // Counter'ı MAX'a zorla (field private → test için session'ı yeniden kuramayız;
-        // bunun yerine MAX+1 başarısız olana kadar supersede yapmak yerine, doğrudan
-        // checked_add overflow'ı tetiklemek için supersedes'ı MAX'a set eden bir helper yok.
-        // Bu yüzden u32::MAX supersede yapıp bir sonrakinde exhaust edelim — pratik değil.
-        //
-        // Alternatif: counter u32; bir supersede yap, sonra session'ı MAX-1'e getirmek
-        // yerine, test invariant'ı checked_add'in mutation öncesi olduğunu zaten gösterir.
-        // PR #49 audit_seq exhaustion testiyle aynı deseni uygula: audit_seq test helper
-        // store'da var (set_audit_seq_for_tests) ama session counter'da yok.
-        //
-        // Çözüm: counter overflow'u u32::MAX'ta tetikle — session.supersedes private field.
-        // Test modülü içinde olduğumuz için field erişimi: session.supersedes = u32::MAX.
-        session.supersedes = u32::MAX;
+        // Test modülü aynı dosyada olduğundan private field'a doğrudan erişilir.
+        session.supersedes = u64::MAX;
 
         let before = snapshot_store(&store);
         let err = session
@@ -2508,6 +2497,6 @@ mod tests {
             "got {err:?}"
         );
         assert_eq!(snapshot_store(&store), before, "store unchanged");
-        assert_eq!(session.supersedes, u32::MAX, "counter unchanged on error");
+        assert_eq!(session.supersedes, u64::MAX, "counter unchanged on error");
     }
 }
